remove defensive checks
