+++ packages/refs/src/index.ts
@@ -1,5 +1,7 @@
 export type { Segment, PathInput, RefOptions, RefContext } from "./types";
 
-export { Ref, ref } from "./ref";
+export { Ref } from "./ref";
 
 export { at, findRef } from "./utils";
+
+export { ref } from "./factory";
